✨ Add Google reCAPTCHA v3 and Cloudflare Turnstile support for chat widget
- 🛡️ Add provider-based captcha validation (Google reCAPTCHA v3 + Cloudflare Turnstile)
- 🔐 Extend captcha settings with back-office/chat-specific enable flags and score threshold
- 🧩 Add dedicated captcha helper module for widget (script loading + token resolution)
- 📦 Wire captcha tokens into online chat submission flow
- 🎨 Add captcha provider toggle UI in admin recaptcha settings
- 🗺️ Expose captcha settings through widget online settings endpoint
- ⚡ Update source maps and bundled assets for new captcha chunk

+++ lhc_web/lib/core/lhgenericbot/actionTypes/lhgenericbotactioncommand.php
@@ -490,7 +490,7 @@ class erLhcoreClassGenericBotActionCommand {
                     $groupedFields = [];
 
                     $messagesGroupFieldAll = isset($action['content']['payload_arg_val_sum']) && !empty($action['content']['payload_arg_val_sum']) ? explode(',',$action['content']['payload_arg_val_sum']) : [];
-                    $messagesThresholdValue = isset($action['content']['payload_arg_val_trshl']) && !empty($action['content']['payload_arg_val_trshl']) ? (double)$action['content']['payload_arg_val_trshl'] : 0;
+                    $messagesThresholdValue = isset($action['content']['payload_arg_val_trshl']) && !empty($action['content']['payload_arg_val_trshl']) ? (float)$action['content']['payload_arg_val_trshl'] : 0;
 
                     foreach ($messages as $message) {
                         $messageVariables = $message->meta_msg_array;
